+++ design-system/react/src/components/Dropdown/Dropdown.test.tsx
@@ -40,9 +40,7 @@ describe('Dropdown', () => {
 
     expect(() => screen.getByText('Settings')).toThrow();
     const trigger = screen.getByText('Click here');
-    await act(async () => {
-      await user.click(trigger);
-    });
+    await act(() => user.click(trigger));
 
     expect(await screen.findByText('Settings')).toBeInTheDocument();
   });
@@ -51,14 +49,10 @@ describe('Dropdown', () => {
     const { user } = render(<Content />);
 
     const trigger = screen.getByText('Click here');
-    await act(async () => {
-      await user.click(trigger);
-    });
+    await act(() => user.click(trigger));
 
     const menuItem = await screen.findByText(/settings/i);
-    await act(async () => {
-      await user.click(menuItem);
-    });
+    await act(() => user.click(menuItem));
 
     await waitFor(() => expect(() => screen.getByText('Settings')).toThrow());
   });
@@ -72,14 +66,10 @@ describe('Dropdown', () => {
     );
 
     const trigger = screen.getByText('Click here');
-    await act(async () => {
-      await user.click(trigger);
-    });
+    await act(() => user.click(trigger));
 
     const fooBtn = await screen.findByText('Foo');
-    await act(async () => {
-      await user.click(fooBtn);
-    });
+    await act(() => user.click(fooBtn));
 
     await waitFor(() => expect(() => screen.getByText('Settings')).toThrow());
   });
@@ -88,9 +78,7 @@ describe('Dropdown', () => {
     const { user } = render(<Content />);
 
     const trigger = screen.getByText('Click here');
-    await act(async () => {
-      await user.click(trigger);
-    });
+    await act(() => user.click(trigger));
     await user.press('Esc');
     await waitFor(() => expect(() => screen.getByText('Settings')).toThrow());
   });