+++ design-system/react/src/components/InputAmount/InputAmount.stories.tsx
@@ -1,14 +1,14 @@
 import type { BN } from '@fuel-ts/math';
 import { bn } from '@fuel-ts/math';
 import { action } from '@storybook/addon-actions';
+import type { ComponentStory, ComponentMeta } from '@storybook/react';
 import { useEffect, useState } from 'react';
 
 import { Box } from '../Box';
 import { Button } from '../Button';
-import { Flex } from '../Flex';
+import { Stack } from '../Stack';
 import { Text } from '../Text';
 
-import type { InputAmountProps } from './InputAmount';
 import { InputAmount } from './InputAmount';
 
 export default {
@@ -24,11 +24,14 @@ export default {
       control: 'boolean',
     },
   },
-};
+} as ComponentMeta<typeof InputAmount>;
 
-const Template = (args: InputAmountProps) => {
+const BALANCE = bn.parseUnits('1.570000044');
+
+const Template: ComponentStory<typeof InputAmount> = (args) => {
   const [amount, setAmount] = useState<BN | null>(bn());
-  const AMOUNT_VALUE = 1_000_000_011;
+  const AMOUNT_VALUE_1 = 1_000_000_011;
+  const AMOUNT_VALUE_2 = 1_000_000_000;
 
   // Log onChange amount
   useEffect(() => {
@@ -39,42 +42,34 @@ const Template = (args: InputAmountProps) => {
 
   return (
     <Box css={{ width: 300 }}>
-      <InputAmount
-        {...args}
-        balance={bn.parseUnits('1.57')}
-        onChange={(e) => {
-          setAmount(e);
-        }}
-        value={amount}
-      />
-      <Text fontSize="lg" css={{ marginTop: '$2' }}>
-        Amount:{' '}
-        {amount?.format({
-          precision: 9,
-        })}
-      </Text>
-      <Flex css={{ gap: '$2', marginTop: '$3' }}>
-        <Button onPress={() => setAmount(bn(AMOUNT_VALUE))}>
-          Set (
-          {bn(AMOUNT_VALUE).format({
-            precision: 9,
-          })}
-          )
+      <InputAmount {...args} onChange={setAmount} value={amount} />
+      <Stack gap="$3">
+        <Text fontSize="lg" css={{ marginTop: '$2' }}>
+          Amount: {amount?.format({ precision: 9 })}
+        </Text>
+        <Button onPress={() => setAmount(bn(AMOUNT_VALUE_1))}>
+          Set ({bn(AMOUNT_VALUE_1).format({ precision: 9 })})
+        </Button>
+        <Button onPress={() => setAmount(bn(AMOUNT_VALUE_2))}>
+          Set ({bn(AMOUNT_VALUE_2).format({ precision: 3 })})
         </Button>
         <Button onPress={() => setAmount(null)}>Clear</Button>
-      </Flex>
+      </Stack>
     </Box>
   );
 };
 
-type TemplateType = typeof Template & {
-  args: Pick<InputAmountProps, 'hiddenBalance' | 'hiddenMaxButton'>;
+export const Usage = Template.bind({});
+Usage.args = {
+  balance: BALANCE,
 };
 
-export const Usage = Template.bind({});
+export const NoBalance = Template.bind({});
+NoBalance.args = {
+  balance: undefined,
+};
 
-export const NoAction = Template.bind({}) as TemplateType;
-
+export const NoAction = Template.bind({});
 NoAction.args = {
   hiddenBalance: true,
   hiddenMaxButton: true,