+++ common/test-utils/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @fuel-ui/test-utils
 
+## 0.12.3
+
+## 0.12.2
+
+## 0.12.1
+
+## 0.12.0
+
+## 0.11.0
+
+### Minor Changes
+
+- ✨ Feat: add new <Pagination> component. Check inside then `Pagination.stories.tsx` to see how to use it, by [@pedronauck](https://github.com/pedronauck) (See [#189](https://github.com/FuelLabs/fuel-ui/pull/189))
+
+## 0.10.2
+
+### Patch Changes
+
+- 🐞 Fix(react): `<InputAmount>` now discount fee when hit max button and also format balance units to don't break layout, by [@pedronauck](https://github.com/pedronauck) (See [#184](https://github.com/FuelLabs/fuel-ui/pull/184))
+
 ## 0.10.1
 
 ## 0.10.0